created a seed.go for users, posts, and comments

+++ internal/model/models.go
@@ -27,9 +27,9 @@ type UserModel struct {
 
 
 type CommentModel struct {
-	ID        int64  `json:"id"`
-	PostID    int64  `json:"post_id"`
-	UserID    int64  `json:"user_id"`
+	ID        int  `json:"id"`
+	PostID    int  `json:"post_id"`
+	UserID    int  `json:"user_id"`
 	Content   string `json:"content"`
 	CreatedAt string `json:"created_at"`
 	User UserModel `json:"user"`
